Maze: fix to make activity start again
The game_name attribute of PyGameActivity seems to be changed
accidentally from 'game' to 'maze'.  As the docstring of
PyGameActivity class states, game_name needs to match the name of a
module that contains the game loop in a function called main.

TIP: please test at least if the activity start before commit :)

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Signed-off-by: Rafael Ortiz <rafael@activitycentral.com>