Allow collaboration (SL #3296)
Replaced "get_buddy" function by "lookup_buddy" as suggested the Error message.

In this version is not allowed to play with more than one player in each XO
(using the gamepad keywords). This is because a known issue [1] and we can not
handle the child's nick properly.

The "mesh.PARTICIPANT_ADD" event sends something like
"6be01ff2bcfaa58eeacc7f10a57b77b65470d413@jabber.sugarlabs.org" in the
"buddy.prosp.nick" variable and the "mesh.broadcast" send the correct child's
name. So, when Maze creates the players uses the mesh.PARTICIPANT_ADD value and
when it reads the messages it find the correct name (the name that the child
chose)

[1] http://dev.laptop.org/ticket/10750

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>
Signed-off-by: Rafael Ortiz <rafael@activitycentral.com>