Collaboration: allow multiple users in the same XO
This is a workaround to the problem with the real nick on a jabber server[1].
Now, we are using the unique identifier (mesh.my_handle or event.handle) for
each player in the game plus a "-%d" at the end of the uid. This value
(Player.uid) is added only when the Activity is shared otherwise this
field is None.

This new workaround allow to distinguish between many users (up to 3) for each
XO laptop -using the gamepad keys.

[1] http://dev.laptop.org/ticket/10750

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>
Signed-off-by: Rafael Ortiz <rafael@activitycentral.com>