Collaboration fixed SL #3747
A bug regarding collaboration was intruced in this commit[1] when the
toolbar was modified to show the correct icons. In the same commit the
call to 'olpcgames.PygameActivity.build_toolbar' was removed and the
collaboration setup was done there.

I copied the code that setup the collaboration from
'olpcgames.PygameActivity.build_toolbar' to
'activity.MazeActivity.__init__'.

[1] http://git.sugarlabs.org/maze/mainline/commit/87f832850ba242b1606acb0f1d60bc5631920f34

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>