Regenerate on rotate
When one laptop in a group rotates or resizes the screen, they may begin
playing with a different maze to the rest of the group.

Now, when a resize event occurs, the new maze will be sent to the group
as if it was selected using easier and harder buttons.  On unrotated
laptops, the new maze is obviously 90 degrees out of aspect ratio.

When others in the group rotate to catch up, a new maze will be
generated each time.

Collaboration between mixed laptops is possible; aspect ratio and
rotation are now longer an essential feature for good collaboration.

Fixes one of the problems identified in http://dev.laptop.org/ticket/8993