Fix occasional different maze while shared
A different maze was shown to two buddies in a shared activity.

Was caused by the age of maze check using a resolution of one second,
though resolution to microseconds was available.

Change to using microsecond resolution for describing age of maze.

Introduces a change to the protocol; where different versions of Maze
are used, the user with a later version will take precedence in choosing
a maze.

Fixes https://github.com/sugarlabs/maze-activity/issues/20