Minimise redraws
With GTK_DEBUG=updates the maze was redrawn entirely each time a player
moved.  This slowed the activity and cost more battery power.

Port from Pygame may be a cause; the draw signal handler for
Gtk.DrawingArea MazeGame was overloaded with actions that were not
required for drawing the widget.

- use the draw signal handler only for painting the changed area from
  the cached surface, add a redraw function for redrawing the cached
  surface, and ensure that every maze change was marked as dirty,

After this was done, players falling into holes did so properly on the
local instance, but were animated in the wrong position on remote
instances.

Before risk holes were added, each instance was responsible for
animating a player move in the maze.  After risk holes were added, the
moving player instance was responsible for notifying others of falling
into a hole.  The notification did arrive at other instances before the
animation had finished moving to the hole coordinates.

- share the responsibility for detecting a fall, so that animation is
  exactly the same on each instance,

- remove the falling notification message; it is redundant.