Share passed holes
When a maze with some passed holes is received by a newly joining
instance, that user is not shown the holes that have been passed.  When
they fall into the hole, they see themselves back at the start, but
others see them beyond the hole.

Share the list of passed holes when the maze definition is sent.